test/requirements.txt
requests==2.32.3
pytest==8.3.2
python-dotenv==1.0.1
pydantic==2.8.2
pytest-html==4.1.1
pytest-dependency==0.5.1
locust==2.24.0
schemathesis==3.30.2
